+++ packages/search/quick-search/docs/0-intro.tsx
@@ -1,13 +1,10 @@
-// @flow
-import React from 'react';
-import { md, Props, code } from '@atlaskit/docs';
+import * as React from 'react';
+import { md, Example, Props, code } from '@atlaskit/docs';
 
 export default md`
-  # Concepts
-
   This is a barebones quick-search component that can render different types of search results.
   
-  # Example
+  ## Usage
   ${code`import { QuickSearch, ObjectResult, ContainerResult, ResultItemGroup } from '@atlaskit/quick-search';
 
   // Inside a react component with proper state and stuff.. :
@@ -30,6 +27,14 @@ export default md`
     );
   }
   `}
+
+  ${(
+    <Example
+      Component={require('../examples/1-Object-Results').default}
+      title="Objects"
+      source={require('!!raw-loader!../examples/1-Object-Results')}
+    />
+  )}
 
   ${(
     <Props