+++ packages/media/media-card/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/media-image
 
+## 44.1.1
+- [patch] [5de3574](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/5de3574):
+
+  - CustomVideoPlayer is now CustomMediaPlayer and supports audio through type property. Media Viewer now uses custom audio player for audio everywhere except IE11.
+
 ## 44.1.0
 - [minor] [c1ea81c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/c1ea81c):