+++ packages/media/media-viewer/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/media-viewer
 
+## 27.1.0
+- [minor] [5de3574](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/5de3574):
+
+  - CustomVideoPlayer is now CustomMediaPlayer and supports audio through type property. Media Viewer now uses custom audio player for audio everywhere except IE11.
+
 ## 27.0.6
 - [patch] [01697a6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/01697a6):